Moved routes folder inside server folder.
